+++ recommender/engines/rl/inference/rl_inference_component.py
@@ -1,5 +1,5 @@
 # pylint: disable=missing-module-docstring, missing-class-docstring, too-few-public-methods
-# pylint: disable=too-many-instance-attributes, no-self-use
+# pylint: disable=too-many-instance-attributes, no-self-use, line-too-long
 
 from typing import Tuple
 
@@ -12,20 +12,30 @@ from recommender.engines.autoencoders.ml_components.embedder import (
 )
 from recommender.engines.panel_id_to_services_number_mapping import K_TO_PANEL_ID
 from recommender.engines.rl.utils import create_state
-from recommender.engines.base.base_inference_component import BaseInferenceComponent
+from recommender.engines.base.base_inference_component import MLEngineInferenceComponent
 from recommender.engines.rl.ml_components.actor import Actor
 from recommender.engines.rl.ml_components.state_encoder import StateEncoder
 from recommender.engines.rl.ml_components.service_selector import ServiceSelector
 from recommender.models import User, SearchData
 
 
-class RLInferenceComponent(BaseInferenceComponent):
+class RLInferenceComponent(MLEngineInferenceComponent):
+    """
+    Recommender engine that provides logged-in users with personalized recommendations in a given context.
+    """
+
+    engine_name = "RL"
+
     def __init__(self, K: int, exploration: bool = False, act_noise: float = 0.0):
         self.exploration = exploration
         self.act_noise = act_noise
         super().__init__(K)
 
     def _load_models(self) -> None:
+        """It loads model or models needed for recommendations and raise
+        exception if it is not available in the database.
+        """
+
         version = K_TO_PANEL_ID.get(self.K)
         self.actor = Actor.load(version=version)
         self.actor.eval()
@@ -37,9 +47,21 @@ class RLInferenceComponent(BaseInferenceComponent):
         )
         self.service_selector = ServiceSelector(self.service_embedder)
 
-    def _for_logged_user(
+    def _generate_recommendations(
         self, user: User, elastic_services: Tuple[int], search_data: SearchData
     ) -> Tuple[int]:
+        """Generate recommendation for logged user.
+
+        Args:
+            user: user for whom recommendation will be generated.
+            elastic_services: item space from the Marketplace.
+            search_data: search phrase and filters information for narrowing
+             down an item space.
+
+        Returns:
+            service_ids: List of recommended services ids.
+        """
+
         state = create_state(user, elastic_services, search_data)
         state_tensors = self.state_encoder([state])
         weights_tensor = self._get_weights(state_tensors)