+++ tests/endpoints/test_recommendations.py
@@ -3,41 +3,13 @@
 import json
 from copy import deepcopy
 
-from pytest import fixture
-
-
-@fixture
-def recommendation_data():
-    return {
-        "user_id": 1,
-        "unique_id": "5642c351-80fe-44cf-b606-304f2f338122",
-        "timestamp": "2021-03-18T18:49:55.006Z",
-        "visit_id": "202090a4-de4c-4230-acba-6e2931d9e37c",
-        "page_id": "some_page_identifier",
-        "panel_id": "v1",
-        "engine_version": "NCF",
-        "elastic_services": [1, 2, 3],
-        "search_data": {
-            "q": "Cloud GPU",
-            "categories": [1],
-            "geographical_availabilities": ["PL"],
-            "order_type": "open_access",
-            "providers": [1],
-            "rating": "5",
-            "related_platforms": [1],
-            "scientific_domains": [1],
-            "sort": "_score",
-            "target_users": [1],
-        },
-    }
-
 
 def test_recommendations(client, mocker, recommendation_data):
     _inference_component_init_mock = mocker.patch(
-        "recommender.engines.base.base_inference_component.BaseInferenceComponent.__init__"
+        "recommender.engines.base.base_inference_component.MLEngineInferenceComponent.__init__"
     )
     inference_component_call_mock = mocker.patch(
-        "recommender.engines.base.base_inference_component.BaseInferenceComponent.__call__"
+        "recommender.engines.base.base_inference_component.MLEngineInferenceComponent.__call__"
     )
     deserializer_mock = mocker.patch(
         "recommender.services.deserializer.Deserializer.deserialize_recommendation"