+++ tests/services/test_engine_selector.py
@@ -1,24 +1,26 @@
 # pylint: disable-all
 
 import pytest
-from pytest import fixture
 import os
 
-from recommender.services.rec_engine_selector import (
+from recommender.services.engine_selector import (
     get_K,
     get_default_recommendation_alg,
     get_engine_names,
     engine_loader,
     load_engine,
-    NCF_ENGINE_NAME,
-    RL_ENGINE_NAME,
 )
-from recommender.engines.rl.inference.rl_inference_component import RLInferenceComponent
+from recommender.engines.rl.inference.rl_inference_component import (
+    RLInferenceComponent,
+)
 from recommender.engines.ncf.inference.ncf_inference_component import (
     NCFInferenceComponent,
 )
+from recommender.engines.random.inference.random_inference_component import (
+    RandomInferenceComponent,
+)
 from recommender.models.ml_component import MLComponent
-from tests.endpoints.test_recommendations import recommendation_data
+from tests.endpoints.conftest import recommendation_data
 from tests.conftest import (
     generate_users_and_services,
 )
@@ -28,15 +30,6 @@ from recommender.errors import (
     InvalidRecommendationPanelIDError,
     NoSavedMLComponentError,
 )
-
-
-@fixture
-def get_engines():
-    engines = {
-        NCF_ENGINE_NAME: NCFInferenceComponent,
-        RL_ENGINE_NAME: RLInferenceComponent,
-    }
-    return engines
 
 
 def test_get_K(recommendation_data):
@@ -61,24 +54,32 @@ def test_get_K(recommendation_data):
             get_K(recommendation_data)
 
 
-def test_get_default_recommendation_alg():
+def test_get_default_recommendation_alg(get_engines):
     """
     Expected behaviour:
-    DEFAULT_RECOMMENDATION_ALG=RL - return RL
-    DEFAULT_RECOMMENDATION_ALG=NCF - return NCF
+    DEFAULT_RECOMMENDATION_ALG in (RL, rl, Rl, rL) - return RL
+    DEFAULT_RECOMMENDATION_ALG in (NCF, ncf, Ncf...) - return NCF
+    DEFAULT_RECOMMENDATION_ALG in (random, Random, ...) - return random
     Not set or something else - return RL
     """
     os.environ["DEFAULT_RECOMMENDATION_ALG"] = "RL"
-    assert get_default_recommendation_alg() == "RL"
-
+    assert get_default_recommendation_alg(get_engines.keys()) == "RL"
     os.environ["DEFAULT_RECOMMENDATION_ALG"] = "NCF"
-    assert get_default_recommendation_alg() == "NCF"
+    assert get_default_recommendation_alg(get_engines.keys()) == "NCF"
+    os.environ["DEFAULT_RECOMMENDATION_ALG"] = "random"
+    assert get_default_recommendation_alg(get_engines.keys()) == "random"
 
     os.environ["DEFAULT_RECOMMENDATION_ALG"] = "NFC"
-    assert get_default_recommendation_alg() == "RL"
-
+    assert get_default_recommendation_alg(get_engines.keys()) == "RL"
     os.environ["DEFAULT_RECOMMENDATION_ALG"] = ""
-    assert get_default_recommendation_alg() == "RL"
+    assert get_default_recommendation_alg(get_engines.keys()) == "RL"
+
+    os.environ["DEFAULT_RECOMMENDATION_ALG"] = "rl"
+    assert get_default_recommendation_alg(get_engines.keys()) == "RL"
+    os.environ["DEFAULT_RECOMMENDATION_ALG"] = "ncf"
+    assert get_default_recommendation_alg(get_engines.keys()) == "NCF"
+    os.environ["DEFAULT_RECOMMENDATION_ALG"] = "Random"
+    assert get_default_recommendation_alg(get_engines.keys()) == "random"
 
 
 def test_get_engine_names(get_engines):
@@ -90,45 +91,55 @@ def test_get_engine_names(get_engines):
     3) Any engine name that exists.
     """
 
-    engine_from_req = ("RL", "NCF")
-    default_engine = ("RL", "NCF")
+    engine_from_req = ["RL", "NCF", "random"]
+    default_engine = ["RL", "NCF", "random"]
     engines_keys = list(get_engines.keys())
 
     # 1. case - engine_from_req is specified, and engines_keys are the same as the range of engine_from_req
     for eg_req in engine_from_req:
         for eg_def in default_engine:
             engine_names = get_engine_names(eg_req, eg_def, engines_keys)
-            expected_eg_names = ["RL", "NCF"] if eg_req == "RL" else ["NCF", "RL"]
+            expected_eg_names = [eg_req]
+
+            if eg_def != eg_req:
+                expected_eg_names.append(eg_def)
+            list_diff = [eng for eng in engine_from_req if eng not in expected_eg_names]
+            expected_eg_names.extend(list_diff)
+
             assert engine_names == expected_eg_names
 
     # 2. case - engine_from_req is NOT specified, and engines_keys are the same as the range of default_engine
     engine_from_req = None
     for eg_def in default_engine:
         engine_names = get_engine_names(engine_from_req, eg_def, engines_keys)
-        expected_eg_names = ["RL", "NCF"] if eg_def == "RL" else ["NCF", "RL"]
+        expected_eg_names = [eg_def]
+        list_diff = [eng for eng in default_engine if eng not in expected_eg_names]
+        expected_eg_names.extend(list_diff)
         assert engine_names == expected_eg_names
 
     # 3. case - engine_from_req and default_engine are NOT specified
     engine_from_req = None
     default_engine = None
     engine_names = get_engine_names(engine_from_req, default_engine, engines_keys)
-    expected_eg_names = [engines_keys[0], engines_keys[1]]
+    expected_eg_names = [engines_keys[0], engines_keys[1], engines_keys[2]]
     assert engine_names == expected_eg_names
 
     # 4. case - engines_keys includes engines than engine_from_req nad default_engine
-    engine_from_req = ("RL", "NCF")
-    default_engine = ("RL", "NCF")
+    engine_from_req = ("RL", "NCF", "random")
+    default_engine = ("RL", "NCF", "random")
     # Last index of engines_keys, so it is expected to be the last element of returned names
     engines_keys.append("New Engine")
 
     for eg_req in engine_from_req:
         for eg_def in default_engine:
             engine_names = get_engine_names(eg_req, eg_def, engines_keys)
-            expected_eg_names = (
-                ["RL", "NCF", "New Engine"]
-                if eg_req == "RL"
-                else ["NCF", "RL", "New Engine"]
-            )
+            expected_eg_names = [eg_req]
+            if eg_def != eg_req:
+                expected_eg_names.append(eg_def)
+            list_diff = [eng for eng in engine_from_req if eng not in expected_eg_names]
+            expected_eg_names.extend(list_diff)
+            expected_eg_names.append("New Engine")
+
             assert engine_names == expected_eg_names
 
 
@@ -140,9 +151,9 @@ def test_engine_loader(
 ):
     """
     Expected behaviour:
-    If there exists any saved engine, return the engine, else raise NoSavedMLComponentError
+    return proper engine based on the order of engine names, else raise NoSavedMLComponentError
     """
-    engine_names = ["RL", "NCF"]
+    engine_names = ["RL", "NCF", "random"]
     engines = get_engines
     K = 3
 
@@ -151,23 +162,38 @@ def test_engine_loader(
     assert type(engine) == RLInferenceComponent
     assert engine_name == "RL"
 
-    engine_names = ["NCF", "RL"]
+    engine_names = ["NCF", "RL", "random"]
     engine, engine_name = engine_loader(engine_names, engines, K)
     assert type(engine) == NCFInferenceComponent
     assert engine_name == "NCF"
 
-    # 2. case no engine is saved
-    MLComponent.drop_collection()
+    engine_names = ["random", "NCF", "RL"]
+    engine, engine_name = engine_loader(engine_names, engines, K)
+    assert type(engine) == RandomInferenceComponent
+    assert engine_name == "random"
 
+    # 2. case no ML engine is saved and random engine is not passed
+    engine_names = ["NCF", "RL"]
+    MLComponent.drop_collection()
     with pytest.raises(NoSavedMLComponentError):
         engine_loader(engine_names, engines, K)
+
+    engine_names = ["RL", "NCF"]
+    with pytest.raises(NoSavedMLComponentError):
+        engine_loader(engine_names, engines, K)
+
+    # 3. case no ML engine is but random engine is also passed
+    engine_names = ["NCF", "RL", "random"]
+    engine, engine_name = engine_loader(engine_names, engines, K)
+    assert type(engine) == RandomInferenceComponent
+    assert engine_name == "random"
 
 
 def test_engine_loader2(
     get_engines, generate_users_and_services, mock_rl_pipeline_exec
 ):
-    # 3. case NCF engine is requested but does not exist - RL engine should be returned
-    engine_names = ["NCF", "RL"]
+    # 4. case NCF engine is requested but does not exist - RL engine should be returned
+    engine_names = ["NCF", "RL", "random"]
     engines = get_engines
     K = 3
 
@@ -181,8 +207,8 @@ def test_engine_loader3(
     generate_users_and_services,
     mock_ncf_pipeline_exec,
 ):
-    # 4. case RL engine is requested but does not exist - NCF engine should be returned
-    engine_names = ["RL", "NCF"]
+    # 5. case RL engine is requested but does not exist - NCF engine should be returned
+    engine_names = ["RL", "NCF", "random"]
     engines = get_engines
     K = 3
 
@@ -199,8 +225,10 @@ def test_load_engine(
 ):
     """
     Expected behaviour:
-    Return engine
+    Return proper engine and its name based on "engine_version"
+    and whether user is logged-in or not
     """
+    # 1. case user is logged-in
     recommendation_data["engine_version"] = "NCF"
     engine, engine_name = load_engine(recommendation_data)
     assert type(engine) == NCFInferenceComponent
@@ -210,3 +238,12 @@ def test_load_engine(
     engine, engine_name = load_engine(recommendation_data)
     assert type(engine) == RLInferenceComponent
     assert engine_name == "RL"
+
+    # 2. user is random
+    del recommendation_data["user_id"]
+
+    for engine_version in {"RL", "NCF", "random", "placeholder"}:
+        recommendation_data["engine_version"] = engine_version
+        engine, engine_name = load_engine(recommendation_data)
+        assert type(engine) == RandomInferenceComponent
+        assert engine_name == "random"